delete pgres

+++ PGReservation/PGReservation/Controllers/PGRegistrationsController.cs
@@ -306,10 +306,21 @@ namespace PGReservation.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            PGRegistration pGRegistration = db.PgRegistrations.Find(id);
-            db.PgRegistrations.Remove(pGRegistration);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            var pgbedslst = db.PgBeds.Where(x => x.PgRegistration.PGID == id)?.Select(y => y.BedID).ToList();
+            if (!db.PgBedPatientInfo.Any(x => pgbedslst.Contains(x.PgBed.BedID)))
+            {
+                db.PgBeds.RemoveRange(db.PgBeds.Where(x => x.PgRegistration.PGID == id));
+                db.SaveChanges();
+                PGRegistration pGRegistration = db.PgRegistrations.Find(id);
+                db.PgRegistrations.Remove(pGRegistration);
+                db.SaveChanges();
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                ModelState.AddModelError("Cannot delete", "Error! Unable to delete the PG. PG is in use");
+                return View(db.PgRegistrations.Find(id));
+            }
         }
 
         protected override void Dispose(bool disposing)
